1. optimize: 代码细节

+++ src/main/java/cn/oyzh/easyshell/fx/ftp/ShellFTPFileTableView.java
@@ -20,7 +20,7 @@ import java.util.List;
  */
 public class ShellFTPFileTableView extends ShellFileTableView<ShellFTPClient, ShellFTPFile> implements FXEventListener {
 
-    private ListChangeListener<ShellFileUploadTask> uploadTaskListener = (ListChangeListener<ShellFileUploadTask>) change -> {
+    private ListChangeListener<ShellFileUploadTask> uploadTaskListener = change -> {
         change.next();
         if (change.wasRemoved()) {
             for (ShellFileUploadTask task : change.getRemoved()) {
@@ -31,7 +31,7 @@ public class ShellFTPFileTableView extends ShellFileTableView<ShellFTPClient, Sh
         }
     };
 
-    private ListChangeListener<ShellFileDeleteTask> deleteTaskListener = (ListChangeListener<ShellFileDeleteTask>) change -> {
+    private ListChangeListener<ShellFileDeleteTask> deleteTaskListener = change -> {
         change.next();
         if (change.wasRemoved()) {
             for (ShellFileDeleteTask task : change.getRemoved()) {
@@ -73,8 +73,10 @@ public class ShellFTPFileTableView extends ShellFileTableView<ShellFTPClient, Sh
 
     @Override
     public void destroy() {
-        this.client.uploadTasks().removeListener(this.uploadTaskListener);
-        this.client.deleteTasks().removeListener(this.deleteTaskListener);
+        if (this.client != null) {
+            this.client.uploadTasks().removeListener(this.uploadTaskListener);
+            this.client.deleteTasks().removeListener(this.deleteTaskListener);
+        }
         this.uploadTaskListener = null;
         this.deleteTaskListener = null;
         super.destroy();
